When a tablet is migrated away, any inactive read which might be reading from said tablet, has to be dropped. Otherwise these inactive reads can prevent sstables from being removed and these sstables can potentially survive until the tablet is migrated back and resurrect data.
This series introduces the fix as well as a reproducer test.

Fixes: https://github.com/scylladb/scylladb/issues/18110

Closes scylladb/scylladb#18179

* github.com:scylladb/scylladb:
  test: add test for cleaning up cached querier on tablet migration
  querier: allow injecting cache entry ttl by error injector
  replica/table: cleanup_tablet(): clear inactive reads for the tablet
  replica/database: introduce clear_inactive_reads_for_tablet()
  replica/database: introduce foreach_reader_concurrency_semaphore
  reader_concurrency_semaphore: add range param to evict_inactive_reads_for_table()
  reader_concurrency_semaphore: allow storing a range with the inactive reader
  reader_concurrency_semaphore: avoid detach() in inactive_read_handle::abandon()

+++ reader_concurrency_semaphore.cc
@@ -28,15 +28,18 @@ logger rcslog("reader_concurrency_semaphore");
 
 struct reader_concurrency_semaphore::inactive_read {
     flat_mutation_reader_v2 reader;
+    const dht::partition_range* range = nullptr;
     eviction_notify_handler notify_handler;
     timer<lowres_clock> ttl_timer;
     inactive_read_handle* handle = nullptr;
 
-    explicit inactive_read(flat_mutation_reader_v2 reader_) noexcept
+    explicit inactive_read(flat_mutation_reader_v2 reader_, const dht::partition_range* range_) noexcept
         : reader(std::move(reader_))
+        , range(range_)
     { }
     inactive_read(inactive_read&& o)
         : reader(std::move(o.reader))
+        , range(o.range)
         , notify_handler(std::move(o.notify_handler))
         , ttl_timer(std::move(o.ttl_timer))
         , handle(o.handle)
@@ -849,6 +852,12 @@ void reader_concurrency_semaphore::inactive_read_handle::abandon() noexcept {
         auto& sem = permit.semaphore();
         sem.close_reader(std::move(permit.aux_data().ir->reader));
         sem.dequeue_permit(permit);
+        // Break the handle <-> inactive read connection, to prevent the inactive
+        // read attempting to detach(). Not only is that unnecessary (the handle
+        // is abandoning the inactive read), but detach() will reset _permit,
+        // which might be the last permit instance alive. Destroying it could
+        // yank out *this from under our feet.
+        permit.aux_data().ir->handle = nullptr;
         permit.aux_data().ir.reset();
     }
 }
@@ -1063,7 +1072,8 @@ reader_concurrency_semaphore::~reader_concurrency_semaphore() {
     }
 }
 
-reader_concurrency_semaphore::inactive_read_handle reader_concurrency_semaphore::register_inactive_read(flat_mutation_reader_v2 reader) noexcept {
+reader_concurrency_semaphore::inactive_read_handle reader_concurrency_semaphore::register_inactive_read(flat_mutation_reader_v2 reader,
+        const dht::partition_range* range) noexcept {
     auto& permit = reader.permit();
     if (permit->get_state() == reader_permit::state::waiting_for_memory) {
         // Kill all outstanding memory requests, the read is going to be evicted.
@@ -1077,7 +1087,7 @@ reader_concurrency_semaphore::inactive_read_handle reader_concurrency_semaphore:
     }
     if (!should_evict_inactive_read()) {
       try {
-        permit->aux_data().ir.emplace(std::move(reader));
+        permit->aux_data().ir.emplace(std::move(reader), range);
         permit->unlink();
         _inactive_reads.push_back(*permit);
         ++_stats.inactive_reads;
@@ -1151,14 +1161,21 @@ void reader_concurrency_semaphore::clear_inactive_reads() {
     }
 }
 
-future<> reader_concurrency_semaphore::evict_inactive_reads_for_table(table_id id) noexcept {
+future<> reader_concurrency_semaphore::evict_inactive_reads_for_table(table_id id, const dht::partition_range* range) noexcept {
+    auto overlaps_with_range = [range] (const reader_concurrency_semaphore::inactive_read& ir) {
+        if (!range || !ir.range) {
+            return true;
+        }
+        return ir.range->overlaps(*range, dht::ring_position_comparator(*ir.reader.schema()));
+    };
+
     permit_list_type evicted_readers;
     auto it = _inactive_reads.begin();
     while (it != _inactive_reads.end()) {
         auto& permit = *it;
         auto& ir = *permit.aux_data().ir;
         ++it;
-        if (ir.reader.schema()->id() == id) {
+        if (ir.reader.schema()->id() == id && overlaps_with_range(ir)) {
             do_detach_inactive_reader(permit, evict_reason::manual);
             permit.unlink();
             evicted_readers.push_back(permit);
